[REFACTOR/#12] 컴포넌트 분리 및 이름 수정
