+++ feature/src/main/java/com/terning/feature/home/component/HomeFilteringScreen.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.home
+package com.terning.feature.home.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
